requirements.txt
pandas>=2.0.0
numpy>=1.23.0
yfinance>=0.2.36
matplotlib>=3.7.0
PyYAML>=6.0.0
